pyproject.toml
[project]
name = "antigravity-jarvis"
version = "0.1.0"
description = "Offline desktop AI companion with local LLMs, voice, memory, automation, and 3D avatar integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1",
  "rich>=13.7.1",
]

[project.optional-dependencies]
voice = [
  "faster-whisper>=1.0.3",
  "openwakeword>=0.6.0",
  "sounddevice>=0.4.7",
]
tts = [
  "piper-tts>=1.2.0",
]
memory = [
  "chromadb>=0.5.0",
  "sentence-transformers>=2.7.0",
]
desktop = [
  "PySide6>=6.7.0",
  "qasync>=0.27.1",
]
local-llm = [
  "httpx>=0.27.0",
]
automation = [
  "apscheduler>=3.10.4",
  "psutil>=5.9.8",
]

[project.scripts]
jarvis = "antigravity_jarvis.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
